+++ docs/RELEASE-NOTES.md
@@ -9,6 +9,7 @@
     - has an update manual
   - deprecated_features:
     - synonym `that` is no longer supported
+    - synonym `equals` is no longer supported
   - bug_fixes:
     - fixes translation of tuple definition
     - fixes translation of synonyms at the beginning of a line